Re-add the pull-request template
But in a very cut-down form.  We don't have an issue tracker on this
repo, and we assume the intentions and motivations for any PR will
have been discussed within the team already.  We do, however, want all
PRs to be against the 'develop' branch, in parallel to upstream's
workflow.