Develop (#9)
* Update dependency dulwich to v0.21.6

* Update dependency django-storages to v1.14

* Update actions/checkout action to v4

* Update Nginx unit to 1.31

* Update docker/setup-buildx-action action to v3

* Update docker/setup-qemu-action action to v3

* Update docker/login-action action to v3

* Update docker.io/postgres Docker tag to v16

* Update dependency psycopg to v3.1.11

* Update postgres Docker tag to v16

* Update dependency psycopg to v3.1.12

* Update dependency django-storages to v1.14.1

* Update dependency django-storages to v1.14.2

* Update dependency django-auth-ldap to v4.6.0

* Update dependency python3-saml to v1.16.0

* Update actions/setup-python action to v5

* Update dependency psycopg to v3.1.16

* Update dependency dulwich to v0.21.7

* Preparation for 2.8.0

* Update README.md (#1)

* Update README.md

* We don't want to use the netbox-community pull-request template

* Update container building workflow

* Only build amd64 images
* Add our 'oxcert' branch as equivalent to upstream's 'release' branch

* Tag the version of the container as OxCERT's modification

This is just for the netbox container: the underlying netbox version
will come from upstream

* Build own own images and push to ghcr.io

* Only build amd64 images
* Only push images to GitHub Container Registry for the OxCERT organization

* Refer to our private copy in the GitHub Container Registry ghcr.io

* Use ghcr.io/oxcert/netbox for all netbox images

Build and push to this repo. Read from it with docker-compose

* Make releases relative to the 'oxcert' branch

rather than upstream's 'release' branch.  This is a different workflow
than used in any of our other repos, where PRs, changes, etc. are
first merged into a 'develop' branch, and the release process is to
merge accumulated changes into the main 'oxcert' branch.

Tag names for releases should follow whatever upstream is using with
"-oxcert" appended.

* Re-add the pull-request template

But in a very cut-down form.  We don't have an issue tracker on this
repo, and we assume the intentions and motivations for any PR will
have been discussed within the team already.  We do, however, want all
PRs to be against the 'develop' branch, in parallel to upstream's
workflow.

* On second thoughts, set container version to 1.0.0

This is OxCERT's version 1.0.0 of the containerized Netbox image which
is basically the same as the 2.7.0 netbox-community equivalent.

* Fix typo from upstream

Function should have been called `git_rebase()` rather than
duplicating the name of `git_merge()`

* Update version numbers

After pulling in changes from upstream

* Reduce the diff between develop and oxcert

We certainly do want to build arm64 images.  Just keep the added
newline on VERSION

* We don't care about building from the 'feature' branch of netbox

This are proposed changes to upstream netbox, which are worked on in a
separate branch called 'feature'.  We only want releases or the latest
netbox code.

Since building from the feature branch is currently failing due to a
permissions problem accessing the docker repo
(https://github.com/netbox-community/netbox.git) this should clean up
our CI actions again.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Tobias Genannt <tobias.genannt@gmail.com>