Don't try and build containers for prerelease versions of netbox (#10)
* Don't try and build containers for prerelease versions of netbox

The netbox_dns plugin is currently incompatible with docker-4.0: see

https://github.com/peteeckel/netbox-plugin-dns/releases/tag/1.0-beta1

So, in the mean time only build our containers pre-seeded with the
netbox_dns plugin for the current relese version(s) of netbox.  We
never want to run the pre-release version of netbox in any case.

* Bump version

Missed then during the previous update.  Needs to be better.