On second thoughts...
Rename to `docker-compose.override.yml` so the overrides are
automatically included in the build.

Initially my thought was that adding the netbox-dns-plugn was causing
failures in the CI testing.  But we're still seeing the same effects
even without the plugin present in the mix, so the conclusion is the
apparent errors are not show stoppers, nor indeed actually errors.