For the time being, don't build arm64 images for testing
We don't have a suitable platform to run the tests on arm64 right now.
Running arm64 code under emulation in QEMU is horribly slow: so slow
that the workflow will timeout and fail.  Native arm64 platforms
aren't available in github at the moment (apparently they are coming
later this year?)  So skip testing on arm64 temporarily.