Clear down previous contents of this repo
The original state of the repo has been preserved in the 'oxcert' branch

This started out as a fork of the netbox-community/netbox-docker
repository on GitHub, and the intention was to apply some local
patches to the source in this repo and then build `netbox-docker`
containers from scratch with those local customizations.

Instead, switch to taking the pre-built docker container from that
project, and apply our changes buy adding layers to that image.

To do that, most of the existing contents of the repository are no
longer needed.  This commit basically deletes almost everything.  The
only content that is left is:

    * LICENSE -- switch to our default MIT License.  This applies to
      OxCERT written code forming the contents of this repository
      only.  The containers we build will use upstream sources with
      their own licensing.  In particular the netbox-docker project
      uses the Apache-2.0 license.

    * README.md -- delete the previous content and add a short
      explanation of how the new repository will work.

    * .yamllint.yaml -- replace with slightly modified settings.

    * .gitignore -- use setting copied from another of our
      repositories which follow the usual OxCERT norms.