Basic setup (#16)
* Clear down previous contents of this repo

The original state of the repo has been preserved in the 'oxcert' branch

This started out as a fork of the netbox-community/netbox-docker
repository on GitHub, and the intention was to apply some local
patches to the source in this repo and then build `netbox-docker`
containers from scratch with those local customizations.

Instead, switch to taking the pre-built docker container from that
project, and apply our changes buy adding layers to that image.

To do that, most of the existing contents of the repository are no
longer needed.  This commit basically deletes almost everything.  The
only content that is left is:

    * LICENSE -- switch to our default MIT License.  This applies to
      OxCERT written code forming the contents of this repository
      only.  The containers we build will use upstream sources with
      their own licensing.  In particular the netbox-docker project
      uses the Apache-2.0 license.

    * README.md -- delete the previous content and add a short
      explanation of how the new repository will work.

    * .yamllint.yaml -- replace with slightly modified settings.

    * .gitignore -- use setting copied from another of our
      repositories which follow the usual OxCERT norms.

* Add pre-commit configuration

Copied from the `debian-systemd` repo.  This should have checks for
all of the interesting bits we're going to add to this repo.

* Add a docker compose specification

Copied from
https://github.com/netbox-community/netbox-docker/blob/release/docker-compose.yml
and only lightly edited.  We're just going to use the latest docker
container from our own ghcr.io repository, which will be build using
the Dockerfile in this repository.

* Disable the hadolint warning about using explicit version numbers

We're passing in the exact container and version we want as an
argument to the `docker build` process.

* Initial commit -- basically just mirror the upstream container

This makes very minimal changes so far.  We'll add local
customizations later.

* oxcert/omega owns everything here

... and gets to approve all the PRs.

* Add a settings file for probot.

Not sure if probot is actually doing anything.  Lets see if it changes
the short destription of the repository.

* Add our usual 'auto author assign' workflow

He who raised it, gets to merge is.  Or otherwise work on it.

* Add github actions to test and build container

Mostly copied from https://github.com/oxcert/debian-systemd

We're really only interested in the exact upstream container image to
pull, not the Debian version installed in the container, so change the
'VERSION' parameter to 'UPSTREAM'

* Apply pre-commit fixes

* Add a reasonable default for the 'UPSTREAM' argument

* Add a link to useful documentation

The official wiki page explaining how to update a Docker image
with various plugins.